Supports desired move-in dates and fraction baths and new UI's and listing bugs

+++ src/main/java/com/orbitekk/shagriha/property/PropertyReader.java
@@ -91,7 +91,7 @@ public class PropertyReader {
                 rs.getObject("pet_count", Integer.class), rs.getBigDecimal("pet_fee"), rs.getBigDecimal("parking_fee"),
                 rs.getBoolean("smoking_included"), rs.getString("stay_type"), rs.getString("bath_type"),
                 rs.getInt("beds"),
-                rs.getInt("baths"), rs.getInt("square_feet"), rs.getString("property_type"),
+                rs.getBigDecimal("baths"), rs.getInt("square_feet"), rs.getString("property_type"),
                 rs.getTimestamp("posted_at").toInstant(), 0, 0, rs.getLong("location_id"),
                 rs.getObject("manager_user_id", java.util.UUID.class), location, manager);
     }
